+++ Configs/.env.client
@@ -4,11 +4,11 @@ PC_COMMENT = 'TEST'
 PASSWORD = 'FF15ItIsGG'
 PathToWS = 'ws://localhost:3010'
 # VPN Params
-VPNUsage = true
+VPNUsage = false
 VPNFailStop = true
 PathToOVPN = "DBs/VPNconfigs"
 PathToEXEVPN = "C:/Program Files/OpenVPN/bin/openvpn-gui.exe"
-PathToWindscribe = "C:\Program Files (x86)\Windscribe\windscribe-cli.exe"
+PathToWindscribe = "C:\Program Files\Windscribe\windscribe-cli.exe"
 CMD_Post_Timeout = 22000
 CMD_WS_Post_Timeout = 6000
 #General Params
